+++ src/MOI_wrapper/variables.jl
@@ -67,7 +67,7 @@ MOI.supports(::Optimizer, ::MOI.VariablePrimalStart, ::Type{VI}) = true
 
 function MOI.set(model::Optimizer, ::MOI.VariablePrimalStart, vi::MOI.VariableIndex, value::Union{Real, Nothing})
     check_inbounds(model, vi)
-    if value == nothing
+    if value === nothing
         value = 0.0
     end
     model.variable_info[vi.value].start = value
@@ -129,7 +129,7 @@ function MOI.add_constraint(model::Optimizer, v::SVF, eq::MOI.EqualTo{Float64})
     vi = v.variable
     check_inbounds(model, vi)
     if isnan(eq.value)
-        @error "Invalid fixed value $(gt.lower)."
+        @error "Invalid fixed value $(eq.value)."
     end
     if has_lower_bound(model, vi)
         @error "Variable $vi has a lower bound. Cannot be fixed."