+++ src/solver.jl
@@ -1,5 +1,3 @@
-export JuniperSolver
-
 #=
 A solver for MINLP problems using a NLP solver and Branch and Bound
 =#
@@ -7,6 +5,7 @@ A solver for MINLP problems using a NLP solver and Branch and Bound
 function get_default_options()
     nl_solver                           = nothing
     log_levels                          = [:Options,:Table,:Info]
+    silent                              = false
     atol                                = 1e-6
     num_resolve_root_relaxation         = 3
     branch_strategy                     = :StrongPseudoCost
@@ -47,6 +46,7 @@ function get_default_options()
     allow_almost_solved_integral        = true
     registered_functions                = nothing
 
+
     # Only for testing
     force_parallel                      = false
     debug                               = false
@@ -55,7 +55,7 @@ function get_default_options()
 
     fixed_gain_mu                       = false
 
-    return SolverOptions(nl_solver,log_levels,atol,num_resolve_root_relaxation,branch_strategy,gain_mu,
+    return SolverOptions(nl_solver,log_levels,silent,atol,num_resolve_root_relaxation,branch_strategy,gain_mu,
         strong_branching_perc,strong_branching_nsteps,strong_branching_approx_time_limit,strong_restart,
         reliability_branching_threshold,reliability_branching_perc,
         incumbent_constr,obj_epsilon,time_limit,mip_gap,best_obj_stop,solution_limit,all_solutions,
@@ -95,7 +95,7 @@ function combine_options(options)
     end
 
     defaults = get_default_options()
-    if defaults.feasibility_pump == true && (!haskey(options_dict, :mip_solver) || options_dict[:mip_solver] == nothing)
+    if defaults.feasibility_pump == true && (!haskey(options_dict, :mip_solver) || options_dict[:mip_solver] === nothing)
         defaults.feasibility_pump = false
     end
 
@@ -115,7 +115,7 @@ function combine_options(options)
         if haskey(options_dict, fname)
             # check that mip_solver is defined if feasibile pump should be used
             if fname == :feasibility_pump && options_dict[:feasibility_pump] == true
-                if !haskey(options_dict, :mip_solver) || options_dict[:mip_solver] == nothing
+                if !haskey(options_dict, :mip_solver) || options_dict[:mip_solver] === nothing
                     @warn "The feasibility pump can only be used if a mip solver is defined."
                     options_dict[:feasibility_pump] = false
                 end