+++ src/table_log.jl
@@ -46,7 +46,7 @@ function is_table_diff(fields, last_arr, new_arr)
     end    
 
     time_idx = findfirst(fields .== :Time)
-    if time_idx != nothing
+    if time_idx !== nothing
         last_arr = vcat(last_arr[1:time_idx-1],last_arr[time_idx+1:end])
         new_arr = vcat(new_arr[1:time_idx-1],new_arr[time_idx+1:end])
     end