+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Juniper.jl Changelog
 
+<<<<<<< HEAD
 ### Unreleased
 - Solving the two children of a node in parallel
   - Corresponding options `processors` and `two_processors_per_node` 
+=======
+### 0.5.0
+- Upgraded to MOI 0.9.1 and JuMP v0.20.0
+  - Implementation of `TimeLimitSec`, `Silent` & `RawParameter`
+
+### v0.4.3
+- Bugfix: No bounds on binary variables in Feasibility pump [#143](https://github.com/lanl-ansi/Juniper.jl/issues/143)
+- Bugfix: Feasibility pump if no objective in problem [#145](https://github.com/lanl-ansi/Juniper.jl/issues/145)
+>>>>>>> master
 
 ### v0.4.2
 - Ability to not accept `almost` MOI solver status codes in tree search